Поменял OrderEntity.price:Integer ->OrderEntity.price:Double  в базе данных хранится в формате чччч.дд  - 2 цифры после запятой. Убрал из DAO и сервисов "магические строки" релексивно отбражающие поля Entity для restrictios -> вынес их в утильный класс: 1) "примитивные" поля доступны как константы утильного класса 2) ассоциированнные объекты через вложенный статический класс Association утильного класса

+++ hotel/trunk/src/main/java/ru/yandex/zhmyd/hotel/model/DisplayedOrder.java
@@ -37,7 +37,7 @@ public class DisplayedOrder {
         return order.getId();
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return order.getPrice();
     }
